+++ src/app/Http/Controllers/Api/UserController.php
@@ -22,15 +22,12 @@ class UserController extends Controller
     public function getAllUsers()
     {
         // Implement logic to fetch all users
+        if (!(request()->user()->can('view-users-page'))){
+            return "No Permission!"; 
+        }
 
-        // if (Auth::user() and Bouncer::is(request()->user())->can('view-users')){
-        if (Auth::user()){
-            $users = User::all(); 
-            return $users;
-        }
-        else {
-            return "No permission";
-        }
+        $users = User::all(); 
+        return $users;
     }
 
     public function getUser($id)
@@ -42,6 +39,11 @@ class UserController extends Controller
     public function getUserAttempts($id)
     {
         // Implement logic to fetch user quiz attempts
+
+        if(!(Auth::user() and request()->user()->can('conduct-quizzes'))) {
+            return "No Permission!";
+        }
+
         $attempts = DB::table('user_attempts')
         ->where('user_id', $id)
         ->get();
@@ -52,6 +54,10 @@ class UserController extends Controller
     public function upsertUserAttempts(Request $request)
     {
         // insert if new, update if exist (update behavior and interpretation given an user attempt)
+
+        if(!(Auth::user() and request()->user()->can('conduct-quizzes'))) {
+            return "No Permission!";
+        }
 
         $user_id = $request->route('id') ?? -1;
         $attempt_id = -1;
@@ -230,6 +236,10 @@ class UserController extends Controller
     public function deleteUserAllAttempts(Request $request, $id) {
         // delete all the user attempts specify an user_id
 
+        if (!(Auth::user() and request()->user()->can('delete-quizzes'))) {
+            return "No Permission!";
+        }
+
         UserAttempt::where('user_id', $id)->delete();
         Attempt::where('user_id', $id)->delete();
     }
@@ -253,6 +263,10 @@ class UserController extends Controller
     public function deleteUser(Request $request, $id)
     {
         // Implement logic to delete user
+        if (!(request()->user()->can('delete-users'))){
+            return "No Permission!"; 
+        }
+        
         User::find($id)->delete();
     }
 
@@ -261,6 +275,10 @@ class UserController extends Controller
     */
     public function deleteUserAttempts($id)
     {
+        if (!(Auth::user() and request()->user()->can('delete-quizzes'))) {
+            return "No Permission!";
+        }
+
         $attemptIDs = Attempt::where('user_id','=',$id)->pluck('id')->toArray();
         foreach ($attemptIDs as $key => $value) {
           $attemptQuiz = AttemptQuiz::where('attempt_id','=',$value);
